pyproject.toml
[project]
name = "langgraph-tutorial"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11.0"
dependencies = [
    "ipykernel>=6.30.1",
    "ipython>=9.4.0",
    "langchain>=0.3.27",
    "langchain-openai>=0.3.30",
    "langgraph>=0.6.4",
    "langsmith>=0.4.13",
    "mypy>=1.17.1",
    "python-dotenv>=1.1.1",
    "tiktoken>=0.11.0",
]
